Summary

+++ DS.EFCore.Helper/DS.EFCore.Helper/DbContextExtensions.cs
@@ -9,6 +9,17 @@ namespace DS.EFCore.Helper
 {
     public static class DbContextExtensions
     {
+        /// <summary>
+        /// <para>
+        /// Begins tracking the given (multiple) entities with State = EntityState.Deleted.
+        /// </para>
+        /// <para>
+        /// Entities with State = EntityState.Added will be detached from the DbContext.
+        /// </para>
+        /// </summary>
+        /// <typeparam name="TEntity">Entity class.</typeparam>
+        /// <param name="dbContext">DbContext that will track changes.</param>
+        /// <param name="entities">Entities to remove.</param>
         public static void RemoveUntrackedEntities<TEntity>(this DbContext dbContext, IEnumerable<TEntity> entities) where TEntity : class
         {
             foreach (TEntity entity in entities)
@@ -22,6 +33,17 @@ namespace DS.EFCore.Helper
             dbContext.RemoveRange(entities);
         }
 
+        /// <summary>
+        /// <para>
+        /// Begins tracking the given (multiple) entities with State = EntityState.Modified.
+        /// </para>
+        /// <para>
+        /// Entities with State = EntityState.Detached will be attached to the DbContext.
+        /// </para>
+        /// </summary>
+        /// <typeparam name="TEntity">Entity class.</typeparam>
+        /// <param name="dbContext">DbContext that will track changes.</param>
+        /// <param name="entities">Entities to update.</param>
         public static void UpdateUntrackedEntities<TEntity>(this DbContext dbContext, IEnumerable<TEntity> entities) where TEntity : class
         {
             foreach (TEntity entity in entities)
@@ -35,6 +57,17 @@ namespace DS.EFCore.Helper
             }
         }
 
+        /// <summary>
+        /// <para>
+        /// Begins tracking the given entity with State = EntityState.Deleted.
+        /// </para>
+        /// <para>
+        /// If the entity has State = EntityState.Added, it will be detached from the DbContext.
+        /// </para>
+        /// </summary>
+        /// <typeparam name="TEntity">Entity class.</typeparam>
+        /// <param name="dbContext">DbContext that will track changes.</param>
+        /// <param name="entity">Entity to remove.</param>
         public static void RemoveUntrackedEntity<TEntity>(this DbContext dbContext, TEntity entity) where TEntity : class
         {
             if (dbContext.Entry(entity).State == EntityState.Detached)
@@ -45,6 +78,17 @@ namespace DS.EFCore.Helper
             dbContext.Remove(entity);
         }
 
+        /// <summary>
+        /// <para>
+        /// Begins tracking the given entity with State = EntityState.Modified.
+        /// <para>
+        /// </para>
+        /// If the entity has State = EntityState.Detached, it will be attached to the DbContext.
+        /// </para>
+        /// </summary>
+        /// <typeparam name="TEntity">Entity class.</typeparam>
+        /// <param name="dbContext">DbContext that will track changes..</param>
+        /// <param name="entity">Entity to update..</param>
         public static void UpdateUntrackedEntity<TEntity>(this DbContext dbContext, TEntity entity) where TEntity : class
         {
             if (dbContext.Entry(entity).State == EntityState.Detached)
@@ -55,6 +99,20 @@ namespace DS.EFCore.Helper
             dbContext.Entry(entity).State = EntityState.Modified;
         }
 
+        /// <summary>
+        /// <para>
+        /// Begins tracking the entity that matches the given expression with State = EntityState.Deleted.
+        /// </para>
+        /// <para>
+        /// If the entity has State = EntityState.Added, it will be detached from the DbContext.
+        /// </para>
+        /// <para>
+        /// If no entity is found, an exception will be thrown.
+        /// </para>
+        /// </summary>
+        /// <typeparam name="TEntity">Entity class.</typeparam>
+        /// <param name="dbContext">DbContext that will track changes.</param>
+        /// <param name="filter">Expression that filters the entity to remove.</param>
         public static void RemoveBy<TEntity>(this DbContext dbContext, Expression<Func<TEntity, bool>> filter) where TEntity : class
         {
             TEntity entity = dbContext
@@ -65,6 +123,21 @@ namespace DS.EFCore.Helper
             dbContext.Remove(entity);
         }
 
+        /// <summary>
+        /// <para>
+        /// Begins tracking asynchronously the entity that matches the given expression with State = EntityState.Deleted.
+        /// </para>
+        /// <para>
+        /// If the entity has State = EntityState.Added, it will be detached from the DbContext.
+        /// </para>
+        /// <para>
+        /// If no entity is found, an exception will be thrown.
+        /// </para>
+        /// </summary>
+        /// <typeparam name="TEntity">Entity class.</typeparam>
+        /// <param name="dbContext">DbContext that will track changes.</param>
+        /// <param name="filter">Expression that filters the entity to remove.</param>
+        /// <returns></returns>
         public static async Task RemoveAsyncBy<TEntity>(this DbContext dbContext, Expression<Func<TEntity, bool>> filter) where TEntity : class
         {
             TEntity entity = await dbContext
@@ -75,6 +148,20 @@ namespace DS.EFCore.Helper
             dbContext.Remove(entity);
         }
 
+        /// <summary>
+        /// <para>
+        /// Begins tracking the (multiple) entities that matches the given expression with State = EntityState.Deleted.
+        /// </para>
+        /// <para>
+        /// Entities with State = EntityState.Added will be detached from the DbContext.
+        /// </para>
+        /// <para>
+        /// If no entity is found, nothing will be done.
+        /// </para>
+        /// </summary>
+        /// <typeparam name="TEntity">Entity class.</typeparam>
+        /// <param name="dbContext">DbContext that will track changes.</param>
+        /// <param name="filter">Expression that filters the entities to remove.</param>
         public static void RemoveAll<TEntity>(this DbContext dbContext, Expression<Func<TEntity, bool>> filter) where TEntity : class
         {
             TEntity[] entities = dbContext
@@ -83,9 +170,25 @@ namespace DS.EFCore.Helper
                 .Where(filter)
                 .ToArray();
 
-            dbContext.RemoveRange(entities);
+            if (entities.Any())
+                dbContext.RemoveRange(entities);
         }
 
+        /// <summary>
+        /// <para>
+        /// Begins tracking asynchronously the (multiple) entities that matches the given expression with State = EntityState.Deleted.
+        /// </para>
+        /// <para>
+        /// Entities with State = EntityState.Added will be detached from the DbContext.
+        /// </para>
+        /// <para>
+        /// If no entity is found, nothing will be done.
+        /// </para>
+        /// </summary>
+        /// <typeparam name="TEntity">Entity class.</typeparam>
+        /// <param name="dbContext">DbContext that will track changes.</param>
+        /// <param name="filter">Expression that filters the entities to remove.</param>
+        /// <returns></returns>
         public static async Task RemoveAllAsync<TEntity>(this DbContext dbContext, Expression<Func<TEntity, bool>> filter) where TEntity : class
         {
             TEntity[] entities = await dbContext
